test: 일기 조회 인터셉터 테스트 수정 및 일기 조회 인가 테스트 추가 (#352)
